Renamed test directory
