+++ python/test/test_model_selection.py
@@ -7,7 +7,7 @@ from sklearn.model_selection import RandomizedSearchCV as \
     SklearnRandomizedSearchCV
 from sklearn.model_selection import GridSearchCV as SklearnGridSearchCV
 
-from tests.pyspark_test import PySparkTest
+from test.pyspark_test import PySparkTest
 
 
 # Overwrite the sklearn GridSearch in this suite so that we can run the same
